Refactored. Added division operation

+++ src/learning/vladdubceac/strategy/Operation.java
@@ -1,5 +1,5 @@
 package learning.vladdubceac.strategy;
 
 public interface Operation {
-    long doOperation(int num1, int num2);
+    double doOperation(int num1, int num2);
 }
