+++ src/learning/vladdubceac/strategy/OperationMultiply.java
@@ -2,7 +2,7 @@ package learning.vladdubceac.strategy;
 
 public class OperationMultiply implements Operation {
     @Override
-    public long doOperation(int num1, int num2) {
+    public double doOperation(int num1, int num2) {
         return num1 * num2;
     }
 }
